requirements.txt
tensorflow
numpy
python-whois
beautifulsoup4
tqdm